// src/core/enums/user-role.enum.ts
export enum UserRoles {
  AdminSystem = "AdminSystem",
  UserReal = "UserReal",
  UserLegal = "UserLegal",
  ContractManager = "ContractManager",
  Seller = "Seller",
}

export enum UserRolesNumber {
  AdminSystem = 1,
  UserReal = 2,
  UserLegal = 3,
  ContractManager = 4,
  Seller = 5,
}

export enum UserRolesPersian {
  AdminSystem = "ادمین سیستم",
  UserReal = "کاربر حقیقی",
  UserLegal = "کاربر حقوقی",
  ContractManager = "مسئول امور قرارداد ها",
  Seller = "فروشنده",
}